Make validation compile
