+++ validation/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 
 [dependencies]
 parity-wasm = { version = "0.31", default-features = false }
+memory_units = "0.3.0"
 
 [features]
 default = ["std"]
